pyproject.toml
[project]
name = "echopilot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate==1.10.1",
    "aiofiles==24.1.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.15",
    "aiosignal==1.4.0",
    "altair==5.5.0",
    "annotated-types==0.7.0",
    "antlr4-python3-runtime==4.9.3",
    "anyio==4.10.0",
    "asttokens==3.0.0",
    "async-timeout==4.0.3",
    "attrs==25.3.0",
    "av==15.1.0",
    "backoff==2.2.1",
    "black==25.1.0",
    "blinker==1.9.0",
    "bokeh==3.8.0",
    "brotli==1.1.0",
    "cachetools==6.2.0",
    "certifi==2025.8.3",
    "charset-normalizer==3.4.3",
    "click==8.2.1",
    "comm==0.2.3",
    "contourpy==1.3.2",
    "coverage==7.10.6",
    "cycler==0.12.1",
    "dataclasses-json==0.6.7",
    "debugpy==1.8.17",
    "decorator==5.2.1",
    "diffusers==0.35.1",
    "distro==1.9.0",
    "einops==0.8.1",
    "exceptiongroup==1.3.0",
    "executing==2.2.1",
    "fastapi==0.116.2",
    "ffmpeg==1.4",
    "ffmpy==0.6.1",
    "filelock==3.19.1",
    "flake8==7.3.0",
    "flask==3.1.2",
    "flask-cors==6.0.1",
    "fonttools==4.59.2",
    "frozenlist==1.7.0",
    "fsspec==2025.9.0",
    "gitdb==4.0.12",
    "gitpython==3.1.45",
    "greenlet==3.2.4",
    "h11==0.16.0",
    "h5py==3.14.0",
    "hf-xet==1.1.9",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.1",
    "huggingface-hub==0.34.4",
    "hydra-core==1.3.2",
    "idna==3.10",
    "imageio==2.37.0",
    "imageio-ffmpeg==0.6.0",
    "importlib-metadata==8.7.0",
    "iniconfig==2.1.0",
    "iopath==0.1.10",
    "ipykernel==6.30.1",
    "ipython==8.37.0",
    "itsdangerous==2.2.0",
    "jedi==0.19.2",
    "jinja2==3.1.6",
    "jiter==0.10.0",
    "joblib==1.5.2",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonschema==4.25.1",
    "jsonschema-specifications==2025.9.1",
    "jupyter-client==8.6.3",
    "jupyter-core==5.8.1",
    "kiwisolver==1.4.9",
    "langchain==0.3.27",
    "langchain-community==0.3.29",
    "langchain-core==0.3.75",
    "langchain-openai==0.3.32",
    "langchain-text-splitters==0.3.11",
    "langgraph==0.6.7",
    "langgraph-checkpoint==2.1.1",
    "langgraph-prebuilt==0.6.4",
    "langgraph-sdk==0.2.6",
    "langsmith==0.4.27",
    "lazy-loader==0.4",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.2",
    "marshmallow==3.26.1",
    "matplotlib==3.10.6",
    "matplotlib-inline==0.1.7",
    "mccabe==0.7.0",
    "mdurl==0.1.2",
    "mpmath==1.3.0",
    "multidict==6.6.4",
    "mypy==1.18.1",
    "mypy-extensions==1.1.0",
    "narwhals==2.5.0",
    "nest-asyncio==1.6.0",
    "networkx==3.4.2",
    "numpy==2.2.6",
    "omegaconf==2.3.0",
    "openai==1.107.0",
    "opencv-python==4.12.0.88",
    "orjson==3.11.3",
    "ormsgpack==1.10.0",
    "packaging==25.0",
    "pandas==2.3.2",
    "parso==0.8.5",
    "pathspec==0.12.1",
    "patsy==1.0.1",
    "pexpect==4.9.0",
    "pillow==11.3.0",
    "platformdirs==4.4.0",
    "plotly==6.3.0",
    "pluggy==1.6.0",
    "portalocker==3.2.0",
    "prompt-toolkit==3.0.52",
    "propcache==0.3.2",
    "protobuf==6.32.1",
    "psutil==7.0.0",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "pyarrow==21.0.0",
    "pycodestyle==2.14.0",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.10.1",
    "pydeck==0.9.1",
    "pydicom==3.0.1",
    "pydub==0.25.1",
    "pyflakes==3.4.0",
    "pygments==2.19.2",
    "pyparsing==3.2.3",
    "pytest==8.4.2",
    "pytest-cov==7.0.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "python-multipart==0.0.20",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "pyzmq==27.1.0",
    "referencing==0.36.2",
    "regex==2025.9.1",
    "requests==2.32.5",
    "requests-toolbelt==1.0.0",
    "rich==14.1.0",
    "rpds-py==0.27.1",
    "ruff==0.13.1",
    "safehttpx==0.1.6",
    "safetensors==0.6.2",
    "scikit-image==0.25.2",
    "scikit-learn==1.7.1",
    "scipy==1.15.3",
    "seaborn==0.13.2",
    "semantic-version==2.10.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "smmap==5.0.2",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.43",
    "stack-data==0.6.3",
    "starlette==0.48.0",
    "statsmodels==0.14.5",
    "streamlit==1.49.1",
    "sympy==1.14.0",
    "tenacity==9.1.2",
    "threadpoolctl==3.6.0",
    "tifffile==2025.5.10",
    "tiktoken==0.11.0",
    "timm==1.0.19",
    "tokenizers==0.22.0",
    "toml==0.10.2",
    "tomli==2.2.1",
    "tomlkit==0.13.3",
    "torch==2.8.0",
    "torchaudio==2.8.0",
    "torchdiffeq==0.2.5",
    "torchvision==0.23.0",
    "tornado==6.5.2",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "transformers==4.56.1",
    "typer==0.17.4",
    "typing-extensions==4.15.0",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "uvicorn==0.35.0",
    "watchdog==6.0.0",
    "wcwidth==0.2.14",
    "websockets==15.0.1",
    "werkzeug==3.1.3",
    "xxhash==3.5.0",
    "xyzservices==2025.4.0",
    "yarl==1.20.1",
    "zipp==3.23.0",
    "zstandard==0.24.0",
]
